move ServerConfig to sozu_command_lib::config
