+++ command/src/config.rs
@@ -1743,6 +1743,47 @@ fn display_toml_error(file: &str, error: &toml::de::Error) {
     }
 }
 
+/// Used by a worker to start its server loop
+pub struct ServerConfig {
+    pub max_connections: usize,
+    pub front_timeout: u32,
+    pub back_timeout: u32,
+    pub connect_timeout: u32,
+    pub zombie_check_interval: u32,
+    pub accept_queue_timeout: u32,
+}
+
+impl ServerConfig {
+    pub fn from_config(config: &Config) -> ServerConfig {
+        ServerConfig {
+            max_connections: config.max_connections,
+            front_timeout: config.front_timeout,
+            back_timeout: config.back_timeout,
+            connect_timeout: config.connect_timeout,
+            zombie_check_interval: config.zombie_check_interval,
+            accept_queue_timeout: config.accept_queue_timeout,
+        }
+    }
+
+    /// size of the slab for the Session manager
+    pub fn slab_capacity(&self) -> usize {
+        10 + 2 * self.max_connections
+    }
+}
+
+impl Default for ServerConfig {
+    fn default() -> ServerConfig {
+        ServerConfig {
+            max_connections: 10000,
+            front_timeout: 60,
+            back_timeout: 30,
+            connect_timeout: 3,
+            zombie_check_interval: 30 * 60,
+            accept_queue_timeout: 60,
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
